Upgrading Messages Plugin
Changed the action names to be more useful. (ie. inbox, read, send)

Made it so that we have a serialized array in the database of everyone who has read the message, and in afterFind() we update the is_read field to be 1 if the current user appears in that array, and 0 if they do not.

Made it so that when you read a message it adds your username to the array of readers and saves it to the database.