Cleanup and Messages Upgrade
Did some cleanup regarding plugin loading and the use of components and behaviors in the messages plugin.  But mainly this is a pretty significant upgrade to the messages plugin and how you interact with messages.  Its coming closer to a more email like, private messaging system, like it is meant to be.