Major Overhaul of the Messages Plugin.
RE-did the schema to make a true messaging system.  The previous schema was making it so that if one user read a message all users read a message. We now have individual labels for each message per each user.  A lot of new functions added, or redone, and unit testing put in. Major overhaul.